Added Boo attack, fixed movement issue
Ghost's boo trick now works. Ghost's hitbox is temporarily removed when trick is active so that they aren't hit by their own attack.

For some reason, Ghost's movement was 1 when it was supposed to be 1.3. Fixed that by specifying move speed in the ghost script, rather than using the parent class's default (which is set to 1.3...)