Disabled AI; updated candy drop mechanic
Game now checks if spawned candy is too close to the targeted player. If so, then candy's position is re-rolled. Also, candy can now spread on both sides of the target, and the range of the spread is reduced.

I disabled AI for the time being. May instead try to get 4 players working on one screen.